fixed the backend for the myWatchlist page so it properly has a controller and function to run on GET
